+++ src/child components/PhotoFrame.jsx
